Add tests

+++ src/main/java/com/nanterre/LoveMyPet/repository/AnimalPerduRepository.java
@@ -4,7 +4,7 @@ import com.nanterre.LoveMyPet.model.AnimalPerdu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface AnimalPerduRepository extends JpaRepository<AnimalPerdu,Long> {
+public interface AnimalPerduRepository extends JpaRepository<AnimalPerdu,Integer> {
